refactor(desktop): modularize managed-agent runtime
Split process ownership, orphan sweeping, dead-instance reaping, lifecycle
synchronization, and runtime metadata into cohesive runtime submodules. Keep
the existing runtime API through narrow re-exports and remove the temporary
runtime.rs file-size override now that the composition root is below the
standard 1000-line ceiling.
